Explicit cast to (uint32_t)
In this part of the code, fragnum is defined as
    for (fragnum = 0; fragnum * MAXIFRAG < buflen; fragnum++) {
so in order for the cast to be safe, we need:
    buflen / MAXIFRAG <= UINT32_MAX

buflen is defined earlier in the function as:
    buflen <= SIZE_MAX
and
    buflen = mind->hindexlen + mind->cindexlen + mind->tindexlen;
where each of those 3 values are <= UINT32_MAX.  Since MAXIFRAG is more than 3
(in fact, it is 261120), we know that (buflen/MAXIFRAG) < UINT32_MAX.

+++ tar/multitape/multitape_metaindex.c
@@ -115,7 +115,7 @@ multitape_metaindex_put(STORAGE_W * S, CHUNKS_W * C,
 		fraglen = buflen - fragnum * MAXIFRAG;
 		if (fraglen > MAXIFRAG)
 			fraglen = MAXIFRAG;
-		multitape_metaindex_fragname(hbuf, fragnum, fraghash);
+		multitape_metaindex_fragname(hbuf, (uint32_t)fragnum, fraghash);
 		if (storage_write_file(S, buf + fragnum * MAXIFRAG,
 		    fraglen, 'i', fraghash))
 			goto err1;
